VectorUpdate changed @vjovanov
This is how VectorUpdate should work, to avoid stage time errors, but I cannot figure out why these errors happens. I really tried to fix it, but now I'm out of ideas. It doesn't make sense at all. The value rhs cast to Rep[Double] cannot be used, it's causing an assertion error, I cannot see the reason. The second thing, the environment is updated with the new vector, and if I print the vector in my test, it has the old value. So this things still prevent the execution of nbodies and drive me crazy. Any help appreciated...